sync lualine and stock statusline colors
This sets the stock statusline background colors to the same colors as
the most dominant background colors in lualine. For lualine, the active
skin contrast is slightly reduced (from title/white/black to fg) in
order to make the statusline stand out a little less, the disabled skin
fg color is changed from disabled to text in order to provide good
readability while maintaining way less contrast in a disabled lualine
compared to an active one.