fix: hide vertical split the best we can
Vertical split is hardcoded to (neo)vim so there's no way getting rid of
it short of making changes to the source code and compiling the binary
yourself. But we can do our best hiding it by setting the bar color to
match background so it blends in. If non_current_windows is set then the
background needs to match non current background in order to blend in
even if none of the vertical splits are selected.